SimpleApp/requirements.txt
sanic==21.6.2
mysql-connector-python==8.0.29
jsonschema==4.4.0
PyJWT==2.3.0
sanic-openapi==21.12.0
sanic-cors==1.0.1
jsonschema==4.4.0
termcolor==1.1.0
requests==2.28.2
web3==5.29.0
